Move tests to top-level test directory and add CTest registration
