+++ test/gmm/GMMFit.cpp
