+++ test/test_hash.cpp
